+++ internal/codehosting/gitlab_test.go
@@ -79,7 +79,7 @@ func TestCreateMergeRequest(t *testing.T) {
 
 	mr, err := gitlab.CreateMergeRequest("Test MR", "This is a test MR", "source-branch", "target-branch")
 	assert.NoError(t, err)
-	assert.Equal(t, 1, mr.ID)
+	assert.Equal(t, int64(1), mr.ID)
 	assert.Equal(t, "http://example.com", mr.URL)
 }
